+++ db/migrate/20260714120100_add_operation_progress_to_compaction_runs.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddOperationProgressToCompactionRuns < ActiveRecord::Migration[8.0]
+  def change
+    add_reference :compaction_runs, :operation_progress, foreign_key: true
+  end
+end